fix dtype

+++ pandas/tests/test_algos.py
@@ -333,10 +333,10 @@ class TestFactorize(object):
         labels, uniques = algos.factorize(data, sort=sort,
                                           na_sentinel=na_sentinel)
         if sort:
-            expected_labels = np.array([1, 0, na_sentinel, 1])
+            expected_labels = np.array([1, 0, na_sentinel, 1], dtype=np.intp)
             expected_uniques = np.array(['a', 'b'], dtype=object)
         else:
-            expected_labels = np.array([0, 1, na_sentinel, 0])
+            expected_labels = np.array([0, 1, na_sentinel, 0], dtype=np.intp)
             expected_uniques = np.array(['b', 'a'], dtype=object)
         tm.assert_numpy_array_equal(labels, expected_labels)
         tm.assert_numpy_array_equal(uniques, expected_uniques)
